set up for multi lr experiment

+++ src/train_ae.py
@@ -25,7 +25,7 @@ parser.add_argument('--network', type=str, default="AE")
 parser.add_argument('--lr1', type=float, default=0.0001)
 parser.add_argument('--lr2', type=float, default=0.0001)
 parser.add_argument('--lr3', type=float, default=0.0001)
-parser.add_argument('--temperature', type=int, default = 100)
+parser.add_argument('--temperature', type=float, default = 100)
 parser.add_argument('--normalize', action=argparse.BooleanOptionalAction, default=True)
 parser.add_argument('--weight_decay', type=float, default=0)
